refactor(language/runtime): reify Policy, Rules, Effect, Hooks, World

+++ language/runtime/src/runtime/engine/continuation.rs
@@ -2,21 +2,21 @@ use destack_vm as vm;
 
 /// Native continuation handle for event loop integration.
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-pub struct NativeContinuation(u64);
+pub struct NativeContinuation(usize);
 
 impl NativeContinuation {
     /// Create a new native continuation handle.
-    pub const fn new(value: u64) -> Self {
+    pub const fn new(value: usize) -> Self {
         Self(value)
     }
 
     /// Return the raw continuation handle value.
-    pub const fn get(self) -> u64 {
+    pub const fn get(self) -> usize {
         self.0
     }
 }
 
-/// Runnable continuation owned by the runtime engine.
+/// Runnable continuation owned by one agent engine.
 #[derive(Debug)]
 #[allow(clippy::large_enum_variant)] // NOTE #Performance #Cleanup: revisit large runtime continuation payloads
 pub enum EngineContinuation {
